% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{FromKomnrCountyname}
\alias{FromKomnrCountyname}
\title{Get information based on municipality number.}
\usage{
FromKomnrCountyname(komnr_string)
}
\arguments{
\item{komnr_string}{municipality number.}
}
\value{
A string
}
\description{
Get information based on municipality number.
}
\examples{
FromKomnrCountyname('1205')

}
