% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{FromKomnrCountyname_single}
\alias{FromKomnrCountyname_single}
\title{Get information based on municipality number.}
\usage{
FromKomnrCountyname_single(komnr_string)
}
\arguments{
\item{komnr_string}{municipality number.}
}
\value{
A string
}
\description{
Get information based on municipality number.
}
\examples{
FromKomnrCountyname_single('1205')

}
