% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{FromKomnrKomnr}
\alias{FromKomnrKomnr}
\title{Get information based on municipality number.}
\usage{
FromKomnrKomnr(komnr_string, type = "countynr")
}
\arguments{
\item{komnr_string}{municipality number.}
}
\value{
An integer
}
\description{
Get information based on municipality number.
}
\examples{
FromKomnrKomnr('1205')

}
