% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{GetKomnrInfo}
\alias{GetKomnrInfo}
\title{Get informatino based on municipality number.}
\usage{
GetKomnrInfo(komnr_string)
}
\arguments{
\item{komnr_string}{municipality number.}
}
\value{
A list.
}
\description{
Get informatino based on municipality number.
}
\examples{
GetKomnrInfo('1205')

}
